+++ src/screen/NFTCollectionDetailsScreen/Activities/index.tsx
