+++ src/screen/NFTCollectionDetailsScreen/Activities/styles.tsx
