+++ src/screen/NFTCollectionDetailsScreen/Banner/index.tsx
