+++ src/screen/NFTCollectionDetailsScreen/styles.tsx
@@ -1,11 +1,10 @@
 import { Dimensions, StyleSheet } from 'react-native';
 import { getPhonePaddingBottom, getPhonePaddingTop } from '../../util/platform';
-import { TABBAR_HEIGHT } from '../../component/CustomBottomTab';
 
 export const styles = StyleSheet.create({
   container: {
     width: '100%',
-    height: Dimensions.get('window').height - TABBAR_HEIGHT,
+    height: Dimensions.get('window').height,
     paddingTop: getPhonePaddingTop(),
     paddingBottom: getPhonePaddingBottom(),
     overflow: 'scroll',
