+++ src/screen/NFTDetailsScreen/NFTDetails/BasicInfo.tsx
