+++ src/screen/NFTDetailsScreen/NFTDetails/Traits.tsx
