+++ src/screen/NFTDetailsScreen/NFTDetails/index.tsx
