+++ src/screen/NFTDetailsScreen/styles.tsx
