+++ src/screen/WalletScreen/NFTTab.tsx
@@ -70,7 +70,7 @@ const NFTTab = () => {
                 <TouchableOpacity
                   key={item.id}
                   style={{ width: 111, height: 111 }}
-                  onPress={() => navigation.navigate('NFTCollections')}>
+                  onPress={() => navigation.navigate('NFTCollection')}>
                   <Image
                     source={{ uri: item.image }}
                     style={{
